разбили объекты недвижимости по модулям. Первоначальный вариант
