+++ index.html
@@ -250,11 +250,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-<<<<<<< HEAD
-<script src='./js/main.js'></script>
-=======
-  <script src='js/main.js'></script>
->>>>>>> eaf5e074387db670549226e05d39ab48019d18aa
+<script type = 'module' src='./js/main.js'></script>
 </body>
 
 </html>
